Move remaining components from containers to components.
