+++ src/web/components/DepthContourPlot3D/DepthColorMap.ts
