+++ src/web/components/DepthContourPlot3D/colorData.json
