+++ src/web/components/LayerSwitcherMenu/LayerSwitcherMenu.less
