+++ src/web/components/LayerSwitcherMenu/__tests__/LayerSwitcherMenu.test.tsx
