+++ src/web/components/MissionsPanel/MissionsList/MissionsList.less
