+++ src/web/components/MissionsPanel/MissionsPanel.less
