--- src/web/containers/JaiaAbout/JaiaAbout.less
@@ -1,69 +1,0 @@
-#jaia-about-container {
-    width: 300px;
-    background-color: #ffffff;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-
-    display: grid;
-    grid-template-columns: 25px 1fr;
-    grid-template-rows: min-content;
-    padding: 15px;
-    border-radius: 3px;
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-    z-index: 101;
-}
-
-.jaia-about-close-btn-container {
-    all: unset;
-    grid-column: 2 / 3;
-    grid-row: 1 / 2;
-    justify-self: end;
-}
-
-.jaia-about-close-btn-container:hover {
-    cursor: pointer;
-}
-
-.jaia-about-logo {
-    width: auto;
-    height: 90px;
-    grid-column: 1 / 3;
-    grid-row: 2 / 3;
-    justify-self: center;
-}
-
-.jaia-about-contact-container {
-    display: grid;
-    grid-template-columns: 25px 1fr;
-    grid-template-rows: min-content;
-
-    margin-bottom: 12px;
-    padding: 6px;
-    background-color: #e4e4e7;
-    border: 1px solid #d4d4d8;
-    border-radius: 3px;
-
-    grid-column: 1 / 3;
-}
-
-.jaia-about-contact-container.last-jaia-info-container {
-    margin-bottom: 0px;
-}
-
-.jaia-about-label {
-    grid-column: 1 / 3;
-    margin-bottom: 6px;
-    font-size: 1.2rem;
-    font-weight: 600;
-}
-
-.jaia-about-input {
-    grid-column: 2 / 3;
-    font-size: 1.1rem;
-}
-
-.jaia-about-link:visited {
-    color: #000000;
-}
